% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saom_to_df.R
\name{sienaBayesResults}
\alias{sienaBayesResults}
\title{Converts sienaBayesFit objects to a Data.frames for convenient Storing of results}
\usage{
sienaBayesResults(
  sienaBayesFitObject,
  writeToExcel = F,
  writePath = NULL,
  writeName = NULL
)
}
\arguments{
\item{sienaBayesFitObject}{The sienaBayesFit object to utilize}

\item{writeToExcel}{Logical. Should the data.frame also be stored to an Excel File? Defaults to False}

\item{writePath}{Character. Path where the Excel table is going to be stored. Defaults to NULL, in which case the table is stored in the current working directory (if writeToExcel == T)}

\item{writeName}{Character. Name of the Excel table. Defaults to NULL, in which case the table will be named "model1.xlsx" (if writeToExcel == T). Table name does not have to include ".xlsx" -ending}
}
\value{
A Data.Frame of the Name, Estimate, Posterior SE, lower and upper credibility interval limits, t-Value (Estimate / Posterior SE) and significance value (abs(t-Value) > 1.96?)
}
\description{
Converts sienaBayesFit objects to a Data.frames for convenient Storing of results
}
