% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSiena_vcov.R
\name{sienaBayesvcov}
\alias{sienaBayesvcov}
\title{Calculates posterior covariance matrix  of sienaBayesFit parameters}
\usage{
sienaBayesvcov(sienaBayesFitObject, excludeRate = T)
}
\arguments{
\item{sienaBayesFitObject}{The sienaBayesFit object to utilize}

\item{excludeRate}{Whether rate parameters should be excluded, defaults to true}
}
\value{
posterior covariance matrix
}
\description{
Calculates posterior covariance matrix  of sienaBayesFit parameters
}
